backend/requirements.txt
fastapi
uvicorn
asyncpg
pytz
python-dotenv
langchain
openai
